Time: 4 ms (42.62%) | Memory: 69.4 MB (79.55%) - LeetSync

+++ 643-maximum-average-subarray-i/maximum-average-subarray-i.java
@@ -6,7 +6,7 @@ class Solution {
         {
             sum+=nums[i];
         }
-        int max=sum;
+        double max=sum;
 
         for(int i=k;i<n;i++)
         {
